Improve hook environment handling
Avoid modifying the current environment for the process when executing
hooks.  Replace uses of popen(3) and pclose(3) with posix_spawn(3) and
waitpid(2) via simple hook_open() and hook_close() functions.  This
allows us to construct a new set of environment variables for the hook
processes based on the parent environment merged with the configured
environment variables for the export.

The use of posix_spawn(3) also allows us to ensure unwanted file
descriptors are closed in the child process.  Only the stdin and stdout
pipe and whatever stderr is are left open.